feat: implement new theme configuration functions and utilities
