+++ packages/tw-theme/src/functions/generateThemeProperties.ts
@@ -2,8 +2,18 @@ import {ThemeColors } from '../types.ts';
 
 import {
   cssVarName,
-} from '../utilities/index.ts';
+} from '../utilities/colors.ts';
 
+/**
+ * Generates CSS custom properties from a theme object.
+ *
+ * @param theme - A theme object containing color definitions and their values.
+ * @returns A record where keys are CSS custom property names (--color-*) and values are color values.
+ *
+ * @example
+ * const theme = { 'primary': '#3b82f6' };
+ * generateThemeProperties(theme); // { '--color-primary': '#3b82f6' }
+ */
 const generateThemeProperties = (
   theme: ThemeColors
 ): Record<string, string> => {
