+++ packages/tw-theme/src/functions/pluginWithOptions.ts
@@ -1,6 +1,13 @@
 import { PluginOptions } from "src/types.ts"
 
-export const plugin = {
+/**
+ * Provides a utility to create Tailwind plugins with options.
+ *
+ * @param pluginFunction - The main plugin function.
+ * @param configFunction - Optional configuration function for the plugin.
+ * @returns A function that combines the plugin handler and configuration.
+ */
+const plugin = {
   withOptions: (pluginFunction: (options: PluginOptions) => void, configFunction: (options: PluginOptions) => void = (options) => ({})) => {
     const optionsFunction = (options: PluginOptions) => {
       const handler = pluginFunction(options)
@@ -11,3 +18,5 @@ export const plugin = {
     return optionsFunction
   },
 }
+
+export default plugin;
