+++ packages/tw-theme/src/theme.ts
@@ -1,11 +1,17 @@
-import { plugin } from './functions/pluginWithOptions.ts'
 import { PluginOptions, ThemeColors , Flags} from './types.ts';
-import generateThemeProperties from './functions/generateThemeProperties.ts';
-import getThemeNameConfig from './functions/getThemeNameConfig.ts';
+import {
+  generateThemeProperties,
+  formatAndCleanPluginConfig,
+  plugin,
+} from './functions/index.ts';
 
 
 /**
- * Extracts theme variables from plugin options and associates them with a theme name.
+ * Extracts CSS custom properties from plugin options and associates them with a theme name.
+ *
+ * @param pluginOptions - The plugin configuration options.
+ * @param themeName - The name of the theme to associate with the variables.
+ * @returns An object containing the theme name and its variables, or null if invalid.
  */
 const extractThemeVariablesFromOptions = ({pluginOptions, themeName}: { pluginOptions: PluginOptions, themeName: string }): { name: string, vars: ThemeColors } | null => {
   const themeVariables: ThemeColors = {};
@@ -22,16 +28,37 @@ const extractThemeVariablesFromOptions = ({pluginOptions, themeName}: { pluginOp
 }
 
 /**
- * Creates the Tailwind plugin for semantic theming.
+ * Parses a theme string to extract its name and flag.
+ *
+ * @param theme - A string containing the theme name and flag.
+ * @returns An object with the theme name and flag, or null if invalid.
+ */
+const getThemeNameConfig = (theme: string): { name: string; flag: Flags } | null => {
+  const [name, flag] = theme.split(' ');
+  if (
+    name && typeof flag === 'string' &&
+    Object.values(Flags).includes(flag as Flags) &&
+    theme.includes(flag)
+  ) {
+    return { name, flag: flag as Flags };
+  }
+  return null;
+};
+
+/**
+ * Creates a Tailwind CSS plugin for semantic theming.
+ *
+ * @returns A configured Tailwind plugin with theme support.
  */
 const createSemanticThemePlugin = () => {
   return plugin.withOptions(function (options: PluginOptions = {}) {
     return function ({ addBase }) {
       const {
-        root = ':root',
-        name = '',
+        root,
+        name,
         ...pluginOptions
-      } = options;
+      } = formatAndCleanPluginConfig(options);
+
       const nameCnf = getThemeNameConfig(name);
       if (!nameCnf) return;
       const customTheme = extractThemeVariablesFromOptions({ themeName: nameCnf.name, pluginOptions });
