+++ packages/tw-theme/src/themes/rootColors.ts
@@ -1,5 +1,10 @@
 import { ThemeColors } from "src/types.ts";
 
+/**
+ * Defines the root color variables shared across all themes.
+ *
+ * These variables include base colors and content colors.
+ */
 const rootColors: ThemeColors = {
     "--color-base-200": "#000000",
     "--color-base-content": "#ffffff"
